refactor(tests): share the both-registries fixture from conftest
test_conflict_instrumentation.py declared its own parametrized `registry`
fixture; a second instrumentation suite in flight declares a byte-identical
one. Moving it to tests/conftest.py gives both a single definition instead
of a third copy.

Collected test ids are unchanged: the shared fixture keeps the "memory" /
"sqlite" param ids the module already used, so nothing about what runs or
how it is named moves.

The nine other `registry` fixtures in the tree keep their local definitions
and shadow this one. Seven of them id their arms "in_memory" rather than
"memory", so adopting the shared fixture would rewrite every one of their
test ids and break any -k filter written against them — a rename worth
making deliberately, not as a side effect of removing a duplicate.

The registry imports sit inside the fixture body rather than at module
level, matching what pytest_configure already does here: conftest.py is
imported before its own collect_ignore_glob guards take effect, so a
top-level ccs import would fail collection outright instead of skipping.
